refactor(cli): physically move notebooklm/zotero/summarize handlers (ARCH-2)
Real cut-and-paste of 19 handlers into cli_notebooklm.py (9), cli_zotero.py (5),
cli_summarize.py (5). cli.py 7686 -> 6787 (957 deletions). cli.py re-imports the
handlers + _sync_cli_dependencies extended for the 3 modules (cli_common already
held the shared helpers from the cli_search batch).

Anti-shortcut verified: real def blocks (9/5/5), the 19 funcs are GONE from
cli.py, no _load_cli_legacy_handlers shim, no circular imports, full suite 3114
green. Mechanical move delegated to Codex (strict brief); Claude verified.
